Slightly adjust scroll handling and also apply to `ModSettingsContainer`
