+++ osu.Game/Overlays/Mods/ModSettingsArea.cs
@@ -54,6 +54,7 @@ namespace osu.Game.Overlays.Mods
                     {
                         RelativeSizeAxes = Axes.Both,
                         ScrollbarOverlapsContent = false,
+                        ClampExtension = 100,
                         Child = modSettingsFlow = new FillFlowContainer
                         {
                             AutoSizeAxes = Axes.X,
@@ -157,9 +158,10 @@ namespace osu.Game.Overlays.Mods
                         new[] { Empty() },
                         new Drawable[]
                         {
-                            new OsuScrollContainer(Direction.Vertical)
+                            new NestedVerticalScrollContainer
                             {
                                 RelativeSizeAxes = Axes.Both,
+                                ClampExtension = 100,
                                 Child = new FillFlowContainer
                                 {
                                     RelativeSizeAxes = Axes.X,
